+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/Icons.kt
@@ -31,6 +31,7 @@ object Icons {
     var IMPLEMENTING_RPC: Icon = loadIcon("implementingRpc.svg")
 
     var PROCEDURE: Icon = loadIcon("procedure.svg")
+    var PROTO_DECOMPILE: Icon = loadIcon("proto_decompile.svg")
 
     private fun loadIcon(name: String): Icon {
         return IconLoader.getIcon("/io/kanro/idea/plugin/protobuf/icon/$name", Icons::class.java)