+++ src/main/kotlin/io/kanro/idea/plugin/protobuf/lang/util/ProtobufPsiFactory.kt
@@ -6,6 +6,7 @@ import com.intellij.psi.PsiWhiteSpace
 import io.kanro.idea.plugin.protobuf.lang.ProtobufFileType
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufFile
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufImportStatement
+import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufStringValue
 import io.kanro.idea.plugin.protobuf.lang.psi.ProtobufTypeName
 import io.kanro.idea.plugin.protobuf.lang.psi.findChild
 import io.kanro.idea.plugin.protobuf.lang.psi.walkChildren
@@ -24,6 +25,13 @@ object ProtobufPsiFactory {
     fun createTypeName(project: Project, text: String): ProtobufTypeName {
         createFile(project, "message Test { optional $text test = 1; }").walkChildren<ProtobufTypeName> {
             return it
+        }
+        throw IllegalStateException("Wrong type name '$text'")
+    }
+
+    fun createStringValue(project: Project, text: String): ProtobufStringValue {
+        createFile(project, "import \"$text\";").walkChildren<ProtobufImportStatement> {
+            return it.stringValue!!
         }
         throw IllegalStateException("Wrong type name '$text'")
     }